# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/foods": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all foods",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Get all foods",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 15,
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.MultipleDataResp-food_Food"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add new food",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Add new food",
                "parameters": [
                    {
                        "description": "Add request",
                        "name": "food",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/food.AddReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-food_AddFood"
                        }
                    }
                }
            }
        },
        "/foods/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get food by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Get food by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-food_Food"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete food by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Delete food by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-food_Food"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update food by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "foods"
                ],
                "summary": "Update food by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "food id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "food",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/food.UpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.SingleDataResp-food_Food"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.MultipleDataResp-food_Food": {
            "type": "object",
            "required": [
                "data",
                "message",
                "pagination"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "pagination": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/api.Pagination"
                        }
                    ],
                    "x-order": "2"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/food.Food"
                    },
                    "x-order": "2"
                }
            }
        },
...
        "api.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pageSize",
                "totalItem",
                "totalPage"
            ],
            "properties": {
                "page": {
                    "type": "integer",
                    "x-order": "1"
                },
                "pageSize": {
                    "type": "integer",
                    "x-order": "2"
                },
                "totalPage": {
                    "type": "integer",
                    "x-order": "3"
                },
                "totalItem": {
                    "type": "integer",
                    "x-order": "4"
                }
            }
        },
        "api.SingleDataResp-food_AddFood": {
            "type": "object",
            "required": [
                "data",
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/food.AddFood"
                        }
                    ],
                    "x-order": "2"
                }
            }
        },
        "api.SingleDataResp-food_Food": {
            "type": "object",
            "required": [
                "data",
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/food.Food"
                        }
                    ],
                    "x-order": "2"
                }
            }
        },
...
        "food.AddFood": {
            "type": "object",
            "required": [
                "createdAt",
                "description",
                "id",
                "images",
                "locationId",
                "name",
                "ownerId",
                "price",
                "review",
                "updatedAt",
                "userId"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "10"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "11"
                },
                "userId": {
                    "type": "string",
                    "x-order": "2"
                },
                "ownerId": {
                    "type": "string",
                    "x-order": "3"
                },
                "locationId": {
                    "type": "string",
                    "x-order": "4"
                },
                "images": {
                    "type": "string",
                    "x-order": "5"
                },
                "name": {
                    "type": "string",
                    "x-order": "6"
                },
                "description": {
                    "type": "string",
                    "x-order": "7"
                },
                "price": {
                    "type": "integer",
                    "x-order": "8"
                },
                "review": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "food.AddReq": {
            "type": "object",
            "required": [
                "description",
                "images",
                "locationId",
                "name",
                "ownerId",
                "price",
                "review",
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string",
                    "x-order": "1"
                },
                "ownerId": {
                    "type": "string",
                    "x-order": "2"
                },
                "locationId": {
                    "type": "string",
                    "x-order": "3"
                },
                "images": {
                    "type": "string",
                    "x-order": "4"
                },
                "name": {
                    "type": "string",
                    "x-order": "5"
                },
                "description": {
                    "type": "string",
                    "x-order": "6"
                },
                "price": {
                    "type": "integer",
                    "x-order": "7"
                },
                "review": {
                    "type": "string",
                    "x-order": "8"
                }
            }
        },
        "food.Food": {
            "type": "object",
            "required": [
                "createdAt",
                "description",
                "id",
                "images",
                "location",
                "name",
                "ownername",
                "price",
                "review",
                "updatedAt",
                "username"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "x-order": "1"
                },
                "createdAt": {
                    "type": "string",
                    "x-order": "10"
                },
                "updatedAt": {
                    "type": "string",
                    "x-order": "11"
                },
                "username": {
                    "type": "string",
                    "x-order": "2"
                },
                "ownername": {
                    "type": "string",
                    "x-order": "3"
                },
                "location": {
                    "type": "string",
                    "x-order": "4"
                },
                "images": {
                    "type": "string",
                    "x-order": "5"
                },
                "name": {
                    "type": "string",
                    "x-order": "6"
                },
                "description": {
                    "type": "string",
                    "x-order": "7"
                },
                "price": {
                    "type": "integer",
                    "x-order": "8"
                },
                "review": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "food.UpdateReq": {
            "type": "object",
            "required": [
                "description",
                "images",
                "locationId",
                "name",
                "ownerId",
                "price",
                "review",
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string",
                    "x-order": "1"
                },
                "ownerId": {
                    "type": "string",
                    "x-order": "2"
                },
                "locationId": {
                    "type": "string",
                    "x-order": "3"
                },
                "images": {
                    "type": "string",
                    "x-order": "4"
                },
                "name": {
                    "type": "string",
                    "x-order": "5"
                },
                "description": {
                    "type": "string",
                    "x-order": "6"
                },
                "price": {
                    "type": "integer",
                    "x-order": "7"
                },
                "review": {
                    "type": "string",
                    "x-order": "8"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}